requirements.txt
pyopengl==3.1.6
pyrr==0.10.3
numpy==1.22.3
glfw==2.5.1
PyQt6~=6.3.0